chore: restructure utils folder
