+++ routes/api/stripe-webhooks.ts
@@ -1,6 +1,6 @@
 // Copyright 2023 the Deno authors. All rights reserved. MIT license.
 import type { Handlers } from "$fresh/server.ts";
-import { stripe } from "@/utils/stripe.ts";
+import { stripe } from "@/utils/payments.ts";
 import { Stripe } from "stripe";
 import {
   getUserIdByStripeCustomerId,
